requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-dotenv>=1.0

# Agent orchestration + LLM integration
langgraph>=0.2
langchain>=0.2
langchain-core>=0.2

# Providers (choose one via env: LLM_PROVIDER=openai|gemini)
langchain-openai>=0.1
langchain-google-genai>=1.0

